requirements.txt
mysql-connector==2.1.3
